fix(engines): restore Rename + Re-probe, delete dead EngineRow.tsx
EngineRow.tsx implemented the Rename (inline) + Re-probe row actions
spec 03 section 9 still documents and the daemon's rename/reprobe
routes still serve, but its render call was silently dropped from
EnginesScreen.tsx during the Phase 2 catalog-card redesign
(ADR-083/084) without the file itself being deleted -- a regression,
not abandoned code (it shipped and rendered from the original MVP
commit onward). Re-wire Rename + Re-probe into CustomEngineCard, the
closest match to EngineRow's single-registered-engine scope, and
delete the now-superseded EngineRow.tsx.
